remove auto-generated client options for data plane
