+++ test/synapse/2019-06-01/azartifacts/zz_generated_bigdatapools_client.go
@@ -27,11 +27,6 @@ type BigDataPoolsClient struct {
 	pl       runtime.Pipeline
 }
 
-// BigDataPoolsClientOptions contains the optional settings for Client.
-type BigDataPoolsClientOptions struct {
-	azcore.ClientOptions
-}
-
 // NewBigDataPoolsClient creates a new instance of BigDataPoolsClient with the specified values.
 // endpoint - The workspace development endpoint, for example https://myworkspace.dev.azuresynapse.net.
 // credential - used to authorize requests. Usually a credential from azidentity.
