+++ test/synapse/2019-06-01/azartifacts/zz_generated_linkedservice_client.go
@@ -27,11 +27,6 @@ type LinkedServiceClient struct {
 	pl       runtime.Pipeline
 }
 
-// LinkedServiceClientOptions contains the optional settings for Client.
-type LinkedServiceClientOptions struct {
-	azcore.ClientOptions
-}
-
 // NewLinkedServiceClient creates a new instance of LinkedServiceClient with the specified values.
 // endpoint - The workspace development endpoint, for example https://myworkspace.dev.azuresynapse.net.
 // credential - used to authorize requests. Usually a credential from azidentity.
