+++ test/synapse/2019-06-01/azartifacts/zz_generated_notebook_client.go
@@ -27,11 +27,6 @@ type NotebookClient struct {
 	pl       runtime.Pipeline
 }
 
-// NotebookClientOptions contains the optional settings for Client.
-type NotebookClientOptions struct {
-	azcore.ClientOptions
-}
-
 // NewNotebookClient creates a new instance of NotebookClient with the specified values.
 // endpoint - The workspace development endpoint, for example https://myworkspace.dev.azuresynapse.net.
 // credential - used to authorize requests. Usually a credential from azidentity.
