+++ test/synapse/2019-06-01/azartifacts/zz_generated_sqlpools_client.go
@@ -27,11 +27,6 @@ type SQLPoolsClient struct {
 	pl       runtime.Pipeline
 }
 
-// SQLPoolsClientOptions contains the optional settings for Client.
-type SQLPoolsClientOptions struct {
-	azcore.ClientOptions
-}
-
 // NewSQLPoolsClient creates a new instance of SQLPoolsClient with the specified values.
 // endpoint - The workspace development endpoint, for example https://myworkspace.dev.azuresynapse.net.
 // credential - used to authorize requests. Usually a credential from azidentity.
