+++ test/synapse/2019-06-01/azartifacts/zz_generated_sqlscript_client.go
@@ -27,11 +27,6 @@ type SQLScriptClient struct {
 	pl       runtime.Pipeline
 }
 
-// SQLScriptClientOptions contains the optional settings for Client.
-type SQLScriptClientOptions struct {
-	azcore.ClientOptions
-}
-
 // NewSQLScriptClient creates a new instance of SQLScriptClient with the specified values.
 // endpoint - The workspace development endpoint, for example https://myworkspace.dev.azuresynapse.net.
 // credential - used to authorize requests. Usually a credential from azidentity.
