+++ test/synapse/2019-06-01/azartifacts/zz_generated_trigger_client.go
@@ -27,11 +27,6 @@ type TriggerClient struct {
 	pl       runtime.Pipeline
 }
 
-// TriggerClientOptions contains the optional settings for Client.
-type TriggerClientOptions struct {
-	azcore.ClientOptions
-}
-
 // NewTriggerClient creates a new instance of TriggerClient with the specified values.
 // endpoint - The workspace development endpoint, for example https://myworkspace.dev.azuresynapse.net.
 // credential - used to authorize requests. Usually a credential from azidentity.
